restructure refinance modal
